+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 # Release notes
 
+## New in git-machete 3.11.3
+
 ## New in git-machete 3.11.2
 
 - fixed: `git machete` now correctly infers remote for fetching of branch when the branch is associated with more than one remote